refactor: restaurants 역할별 분리
