+++ src/restaurants/index.ts
@@ -1,0 +1,3 @@
+export * from "./coord_schema";
+export * from "./restaurants_contract";
+export * from "./restaurants_router";
