+++ src/restaurants/restaurants_contract.ts
@@ -1,22 +1,10 @@
 import { initContract } from "@ts-rest/core";
 import { z } from "../zod_openapi";
+import { coordsSchema, degreeSchema } from "./coord_schema";
 
 const c = initContract();
 
-export const degreeSchema = z.coerce.number();
-
-export const coordsSchema = z.object({
-  lat: degreeSchema
-    .min(-90)
-    .max(90)
-    .openapi({ description: "위도", example: 37.5667 }),
-  lon: degreeSchema
-    .min(-180)
-    .max(180)
-    .openapi({ description: "경도", example: 126.9784 }),
-});
-
-export const restaurantsRouter = c.router(
+export const restaurantsContract = c.router(
   {
     // getManyByRegion: {},
     getMany: {
